Update bright_treasury.md
There is a need to amend the current grant proposal. When we sent our proposal earlier this year, the bounties module was evolving and we decided to include also the sub-bounties mechanism that was planned for development. Unfortunately, the sub-bounties haven’t been implemented yet, which means that we need to exclude the user stories connected to them from the proposal. They were just additional elements and the planned app flow doesn’t have to change, so we just removed the user stories that applied to the sub-bounties only and edited the stories that mentioned them as well. What we propose to be included instead of the sub-bounties is to prepare and deploy to production Milestone 1 and Milestone 2, which already cover the Proposals flow. We feel it might be beneficial to allow users to use the app already, get the feedback and start to promote it within the community as plan to maintain the app beyond the current grant. We would love for the BrightTreasury to become an important part of the community, a living and useful tool that increases the Treasury activities. Going to production earlier might contribute to this goal.