Amend the project milestone tasks
We will need a solid rationale to merge WasmEdge support upstream into the official Substrate repo. So, we changed the order of tasks to complete the performance eval first and then to propose a merge PR. That means shifting the PR task from milestone 2 to 3.

We have also found that the Substrate approach of using host functions requires a lot of changes in the WasmEdge Rust SDK. That has caused a delay and extra engineering work for the project (already completed). We will keep the total project budget the same, but respectfully ask for each milestone payment to stay the same to account for the extra work needed for milestone 2.

Signed-off-by: Michael Yuan <michael@secondstate.io>