Amend milestone #3 delivery timeline
As discussed here:

https://github.com/w3f/Grants-Program/pull/862#issuecomment-1382385409

the milestone #3 task is much greater than we originally envisioned. We need to reconcile changes to the substrate code base since we started and then merge our changes back into the main tree in order to propose an upstream PR and run the most up-to-date performance tests. So, I extended the time needed for this task to 6 months. Thank you.