Spheroid Earth project application
Spheroid Earth is an open visual positioning platform for AR/XR applications and metaverses.

Our initial toolset aims to simplify the work of researchers in the fields of computer vision, photogrammetry, and structure from motion (SfM). Our long-term goal is to build a Kusama/Polkadot parachain that serves as a global-scale visual search engine.