Validator monitoring service Update Milestones
During the development of Milestone 1, we decided to slightly change the order of tasks. The reason lies in the development process during which we decided to more focus on tg bot updates because all the following tasks (that we moved to Milestone 2) require a fully working tg bot. 
The overall number of deliverables remains unchanged.   
As suggested by evaluator @dsm-w3f in the evolution: https://github.com/w3f/Grant-Milestone-Delivery/blob/master/evaluations/validator_monitoring_service_1_dsm-w3f.md we open an amendment request.